feat: open polygons (Curve) + multi-contour glyphs (Circle as Polygon)
Curve and Circle are now Polygon subclasses instead of separate C++
input types, so they get geometry caching, gradient strokes, zIndex,
and effects for free. C++ Circle.cpp/.hpp and Curve.cpp/.hpp deleted
(factory entries + CMakeLists updated).

Polygon gains open=True support (absolute coords, no bbox shift) and
generateRawContours() for multi-contour shapes. Letters are now
rendered as separate contours (earcut-with-holes, no bridge lines),
fixing dot/counter glyphs (i, j, B, O, etc.) and reducing point count
~4x via true quadratic glyph sampling. BezierPath stroke/fill order
reworked (stroke first, fill on top) to fix junction artifacts at tight
glyph corners, with merged close-corner joins and round-join discs.

Mesh indices migrated from uint16 to uint32 throughout (Mesh,
MeshFactory, BezierPath earcut/GeomCache, renderers) to support the
larger vertex counts from dense glyph stress scenes.

New visual tests: curve, text-stroke. Goldens regenerated across the
suite (31 checks green).

+++ include/window/VulkanWidget.hpp
@@ -237,7 +237,7 @@ namespace VC
         std::vector<Mesh>         m_meshes;
         std::vector<MeshDrawInfo> m_meshDrawInfos;
         std::vector<Vertex>       m_vertices;
-        std::vector<uint16_t>     m_indices;
+        std::vector<uint32_t>     m_indices;
         bool                      m_geomDirty = false;
 
         // ── Per-frame partitioned mesh indices ────────────────────────────
